requirements.txt
flask==3.0.2
flask-cors==4.0.0
torch==2.10.0
transformers==4.39.3
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
nltk==3.8.1
matplotlib==3.8.3
seaborn==0.13.2
gunicorn==21.2.0
pypdf==5.4.0
python-docx==1.1.2
gdown==5.2.0
